requirements.txt
boto3>=1.4.0
clint>=0.5.1
python-magic>=0.4.12
tabulate>=0.7.7
tqdm>=4.8.4
